Move ScopedConnection class to its own file
